chore: separate components, refactor, fix: downgrade apps-graphql to 2.x
