+++ react/Promotions.ts
@@ -1,0 +1,3 @@
+import { Promotions } from './components/Promotions'
+
+export default Promotions
