+++ react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sequra-promotions",
-  "version": "0.0.0",
+  "version": "0.0.10",
   "scripts": {
     "test": "vtex-test-tools test"
   },
